moved EmailField to new Fields folder
