+++ app/page.js
@@ -6,7 +6,7 @@ import Footer from "@/components/ui/footer";
 import { Button } from "@/components/ui/button";
 import Testimonials from "@/components/ui/testimonials";
 import { NameField } from "@/components/ui/nameField";
-import { EmailField } from "@/components/ui/emailField";
+import { EmailField } from "@/components/ui/Fields/emailField";
 import { SignUpForm } from "@/components/forms/SignUpForm";
 import Header from "@/components/ui/header";
